@@ -37,13 +37,13 @@
     "test": "grunt test"
   },
   "devDependencies": {
-    "typescript": "1.8.10",
-    "vscode": "0.11.18",
+    "typescript": "2.0.3",
+    "vscode": "1.0.0",
     "grunt": "1.0.1",
     "grunt-cli": "1.2.0",
     "load-grunt-config": "0.19.2",
     "grunt-contrib-clean": "1.0.0",
-    "grunt-shell": "1.3.1"
+    "grunt-shell": "2.0.0"
   },
   "dependencies": {
     "bower": "1.7.9",
