@@ -1,1 +1,1 @@
-/// <reference path="../node_modules/vscode/typings/index.d.ts" />
+/// <reference path="../node_modules/vscode/vscode.d.ts" />
